src/keyword_service/requirements.txt
python-kafka
pydantic
lipsum